refactor(frontend): modularize lora feature
