+++ app/frontend/src/features/lora/components/lora-gallery/LoraCard.vue
@@ -32,7 +32,7 @@
 import { computed, toRef } from 'vue';
 import LoraCardGrid from './LoraCardGrid.vue';
 import LoraCardList from './LoraCardList.vue';
-import { useLoraCardActions } from '@/composables/lora-gallery';
+import { useLoraCardActions } from '@/features/lora/composables';
 import type { LoraGallerySelectionState, LoraListItem, LoraUpdatePayload } from '@/types';
 
 type LoraCardViewModel = {
