+++ app/frontend/src/features/lora/components/lora-gallery/LoraCardList.vue
