+++ app/frontend/src/features/lora/composables/index.ts
@@ -1,0 +1,1 @@
+export * from './lora-gallery';
