+++ app/frontend/src/features/lora/composables/lora-gallery/useLoraGallerySelection.ts
