+++ app/frontend/src/services/index.ts
@@ -1,4 +1,3 @@
 export * from './apiClient';
 export * from './backendClient';
 export * from './system';
-export * from './lora';
