+++ app/frontend/src/stores/index.ts
@@ -1,5 +1,4 @@
 export * from './settings';
 export * from './app';
 export * from './adminMetrics';
-export * from './adapterCatalog';
