+++ tests/vue/LoraGallery.spec.js
@@ -2,8 +2,26 @@ import { describe, it, expect, vi, beforeEach, beforeAll, afterAll } from 'vites
 import { flushPromises, mount } from '@vue/test-utils';
 import { reactive } from 'vue';
 import { createPinia, setActivePinia } from 'pinia';
-import LoraGallery from '@/components/lora-gallery/LoraGallery.vue';
+import LoraGallery from '@/features/lora/components/lora-gallery/LoraGallery.vue';
 import { useSettingsStore } from '@/stores/settings';
+
+vi.mock('@/services', async (importOriginal) => {
+  const actual = await importOriginal();
+  const backendClientMock = {
+    resolve: vi.fn((path = '') => path),
+    requestJson: vi.fn(),
+    getJson: vi.fn(),
+    postJson: vi.fn(),
+    putJson: vi.fn(),
+    delete: vi.fn(),
+    patchJson: vi.fn(),
+    requestBlob: vi.fn(),
+  };
+  return {
+    ...actual,
+    useBackendClient: vi.fn(() => backendClientMock),
+  };
+});
 
 const mocks = vi.hoisted(() => ({
   fetchAdapterListMock: vi.fn(),
@@ -34,6 +52,7 @@ const updateRouteFromPush = async (location = {}) => {
 
 const routerMocks = {
   push: vi.fn(updateRouteFromPush),
+  replace: vi.fn(updateRouteFromPush),
 };
 
 vi.mock('vue-router', () => ({
@@ -80,7 +99,7 @@ vi.mock('@/composables/shared', async (importOriginal) => {
   };
 });
 
-vi.mock('@/services/lora/loraService', async (importOriginal) => {
+vi.mock('@/features/lora/services', async (importOriginal) => {
   const actual = await importOriginal();
   return {
     ...actual,
